refactor(dashboard): rename `run.py` to `app.py`
Sounds better IMO.
